refactor organizations slider to component
