+++ frontend/src/components/landing/LandingHero/index.tsx
@@ -2,33 +2,9 @@ import { Box, Button, Container, Grid, Stack, Typography } from "@mui/material";
 import Hero from "@public/static/landing/hero.jpg";
 import Image from "next/image";
 import Link from "next/link";
-import { useEffect, useRef, useState } from "react";
-import Organizations from "./Organizations";
+import OrganizationsSlider from "./OrganizationsSlider";
 
 const LandingHero: React.FC = () => {
-  const orgsTitleEl = useRef<HTMLElement>();
-
-  const [sliderActiveIndex, setSliderActiveIndex] = useState(0);
-  const [sliderOffsetX, setSliderOffsetX] = useState(0);
-
-  const onActiveIndexChange = (index: number) => {
-    setSliderActiveIndex(index);
-  };
-
-  const updateSliderOffset = () => {
-    if (orgsTitleEl.current) {
-      setSliderOffsetX(orgsTitleEl.current.offsetWidth + orgsTitleEl.current.offsetLeft);
-    }
-  };
-
-  useEffect(() => {
-    updateSliderOffset();
-    window.addEventListener("resize", updateSliderOffset);
-    return () => {
-      window.removeEventListener("resize", updateSliderOffset);
-    };
-  }, []);
-
   return (
     <>
       <Box
@@ -89,37 +65,7 @@ const LandingHero: React.FC = () => {
           </Box>
         </Box>
       </Box>
-      <Box
-        sx={{
-          width: "100vw",
-          py: { xs: 4, md: 6 },
-          px: 1,
-          bgcolor: "background.neutral",
-          borderTop: "1px solid",
-          borderBottom: "1px solid",
-          borderColor: "divider",
-          overflow: "hidden",
-          position: "relative",
-        }}
-      >
-        <Container>
-          <Stack direction="row" minWidth="max-content" alignItems="center">
-            <Box ref={orgsTitleEl}>
-              <Typography
-                variant="h4"
-                sx={{
-                  opacity: sliderActiveIndex == 0 ? 1 : 0,
-                  transition: (theme) => theme.transitions.create("opacity"),
-                }}
-              >
-                Våre <br />
-                Foreninger
-              </Typography>
-            </Box>
-            <Organizations onActiveIndexChange={onActiveIndexChange} offsetX={sliderOffsetX + 48} />
-          </Stack>
-        </Container>
-      </Box>
+      <OrganizationsSlider />
     </>
   );
 };
